Implement faction-specific AI
