+++ Assets/Prefabs/Golden Hand/Unit 1-10/GoldenHandFootMan.prefab
@@ -135,7 +135,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 8785832555143759317}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 862f94809916af942b067ce64b1f5cb1, type: 3}
+  m_Script: {fileID: 11500000, guid: 6b9e5f62c48a4c02a7791eba754654a1, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!61 &7870320153491638533
