--- Assets/Scripts/EnemyAI.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 862f94809916af942b067ce64b1f5cb1
