+++ Assets/Scripts/IUnitAI.cs
@@ -1,0 +1,4 @@
+public interface IUnitAI
+{
+    void DoEnemyTurn();
+}
